+++ NetCoreAPIMySQL.Data/Repositories/Docente/IDocenteRepository.cs
@@ -10,7 +10,7 @@ namespace NetCoreAPIMySQL.Data.Repositories
     public interface IDocenteRepository
     {
         Task<IEnumerable<Docente>> ConsultarDocentes();
-        Task<Docente> ConsultarDocente(int idDocente);
+        Task<Docente> ConsultarDocente(long idDocente);
         Task<bool> RegistrarDocente(Docente docente);
         Task<bool> ModificarDocente(Docente docente);
         Task<bool> EliminarDocente(Docente docente);
